Extract slim job inspect path off the cold first-dispatch graph
job_list/job_get were lazy-importing jobs/service.ts (~500-module
execution graph including run-kody-registry). Move list/get/inspect
reads to jobs/inspect.ts so the first kody.* call in a cold isolate
does not pay that module evaluation cost.
