+++ packages/worker/src/mcp/capabilities/jobs/job-get.ts
@@ -1,6 +1,7 @@
 import { defineDomainCapability } from '#mcp/capabilities/define-domain-capability.ts'
 import { capabilityDomainNames } from '#mcp/capabilities/domain-metadata.ts'
 import { requireMcpUser } from '#mcp/capabilities/meta/require-user.ts'
+import { getJobInspection } from '#worker/jobs/inspect.ts'
 import { listRunRecords } from '#worker/run-records/service.ts'
 import {
 	buildJobInspectionOutput,
@@ -35,7 +36,6 @@ export const jobGetCapability = defineDomainCapability(
 		outputSchema: jobGetOutputSchema,
 		async handler(args, ctx) {
 			const user = requireMcpUser(ctx.callerContext)
-			const { getJobInspection } = await import('#worker/jobs/service.ts')
 			const jobId = resolveJobGetId(args)
 			const inspection = await getJobInspection({
 				env: ctx.env,
